+++ packages/wallets/trezor/CHANGELOG.md
@@ -1,4 +1,10 @@
-# @knackroot/wallet-trezor
+# @thortradekit/wallet-trezor
+
+## 0.0.2
+
+### Patch Changes
+
+- fix
 
 ## 1.2.19